note 3 rename parses.js to parsers.js
